[breaking] Migrate TextFieldsUtils to typescript (click-to-copy)
I was unable to accurately produce the types required for UpdatablePopover. If this is essential (might-be), we'll need to dive deeper into react and react-bootstrap type defs. Those are not easily grokkable.

Estimate an hour of work to [re]fix this. Apologies.
